Makefile: correct text file for paper.pdf
We unfortunately forgot to initialize a variable, which resulted in
our builds to be overly long and unpredictable. In particular,
each time we tried to build paper.pdf, latexmk was called without
any target. As a result, all pdf files were built instead of just
paper.pdf. This patch corrects this problem, which should reduce
build times and avoid us accidentally build IEEE versions.

This reduces the build time of paper.pdf from 13s to 3s on my machine.