Print comments in paper.pdf and create submission.pdf without comments (#43)
When editing a paper we want to see the version with comments. In VSCode,
it is easy to look at the PDF that is created from the main .tex file,
paper.tex in our case. Hence, we make this version the one with comments.

For submission, we always have submission.pdf that will give us a file
without comments and without extended document size.